+++ src/lib/components/widgets/arm/get-joint-positions.svelte
@@ -42,7 +42,8 @@
 			</tr>
 		</thead>
 		<tbody>
-			{#each displayPositions as pos, index (pos)}
+			<!-- joint ids are just the index in the list of values, we need to use index for the key -->
+			{#each displayPositions as pos, index (index)}
 				<tr>
 					<th> {index} </th>
 					<th> {formatNumeric(pos)} </th>