+++ src/lib/components/widgets/camera/picture-in-picture-button.svelte
@@ -1,80 +1,35 @@
 <script lang="ts">
 	import { Button } from '@viamrobotics/prime-core'
-	import { onMount } from 'svelte'
 
 	import ErrorDisplay from '$lib/components/error.svelte'
 
+	import { usePip } from '../../../pip/context.svelte'
+
 	interface Props {
 		resourceName: string
-		mediaStream: MediaStream | null
-		isShowingPip: boolean
+		rate: number | 'live' | false
 	}
 
-	let { resourceName, mediaStream, isShowingPip = $bindable(false) }: Props = $props()
+	const { resourceName, rate }: Props = $props()
 
-	let videoElement = $state.raw<HTMLVideoElement>()
-	let lastErr = $state.raw<Error>()
+	const pip = usePip()
 
-	const onEnter = () => {
-		isShowingPip = true
-	}
-
-	const onLeave = () => {
-		isShowingPip = false
-	}
-
-	onMount(() => {
-		videoElement?.addEventListener('enterpictureinpicture', onEnter)
-		videoElement?.addEventListener('leavepictureinpicture', onLeave)
-
-		return () => {
-			videoElement?.removeEventListener('enterpictureinpicture', onEnter)
-			videoElement?.removeEventListener('leavepictureinpicture', onLeave)
-			if (isShowingPip) {
-				document.exitPictureInPicture()
-			}
+	$effect(() => {
+		if (resourceName === pip.resourceName) {
+			pip.setRate(rate)
 		}
 	})
-
-	$effect.pre(() => {
-		if (videoElement) {
-			videoElement.srcObject = mediaStream
-		}
-	})
-
-	const togglePictureInPicture = async () => {
-		lastErr = undefined
-
-		try {
-			await (isShowingPip
-				? document.exitPictureInPicture()
-				: videoElement?.requestPictureInPicture())
-		} catch (error) {
-			lastErr = error as Error
-		}
-	}
 </script>
 
 <Button
 	icon="picture-in-picture-top-right"
-	onclick={togglePictureInPicture}
+	onclick={() => pip.toggle(resourceName)}
+	progress={pip.readyState === 'loading' ? 'indeterminate' : undefined}
 >
 	Toggle picture-in-picture
 </Button>
 
-<!-- Display a fixed video element on the page but essentially hide it.
-  There must always be a video element visible for PiP. -->
-<video
-	class="fixed right-0 bottom-0 h-px w-px opacity-[0.01]"
-	bind:this={videoElement}
-	muted
-	autoplay
-	controls={false}
-	playsinline
-	aria-label={`${resourceName} picture-in-picture stream`}
-></video>
-
 <ErrorDisplay
 	class="max-w-50"
-	lastError={lastErr}
+	lastError={pip.error}
 />